cleanup public generator API
move fluid_gen_set_default_values(), enum fluid_gen_flags and struct _fluid_gen_t to private header because no other public function seems to need fluid_gen_t
