+++ src/synth/fluid_gen.h
@@ -33,12 +33,34 @@ typedef struct _fluid_gen_info_t {
 	float def;		/* The default value (cfr. fluid_gen_set_default_values()) */
 } fluid_gen_info_t;
 
+/**
+ * SoundFont generator structure.
+ */
+typedef struct _fluid_gen_t
+{
+  unsigned char flags; /**< Is the generator set or not (#fluid_gen_flags) */
+  double val;          /**< The nominal value */
+  double mod;          /**< Change by modulators */
+  double nrpn;         /**< Change by NRPN messages */
+} fluid_gen_t;
+
+/**
+ * Enum value for 'flags' field of #fluid_gen_t (not really flags).
+ */
+enum fluid_gen_flags
+{
+  GEN_UNUSED,		/**< Generator value is not set */
+  GEN_SET,		/**< Generator value is set */
+  GEN_ABS_NRPN		/**< Generator is an absolute value */
+};
+
 #define fluid_gen_set_mod(_gen, _val)  { (_gen)->mod = (double) (_val); }
 #define fluid_gen_set_nrpn(_gen, _val) { (_gen)->nrpn = (double) (_val); }
 
 fluid_real_t fluid_gen_scale(int gen, float value);
 fluid_real_t fluid_gen_scale_nrpn(int gen, int nrpn);
 int fluid_gen_init(fluid_gen_t* gen, fluid_channel_t* channel);
+int fluid_gen_set_default_values(fluid_gen_t* gen);
 
 
 #endif /* _FLUID_GEN_H */
